// vite/src/lib/types/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      daily_cardio_exercises: {
        Row: {
          cardio_name: string;
          created_at: string;
          duration_minutes: number;
          end_time: string;
          id: string;
          start_time: string;
          user_email: string;
          user_id: string | null;
        };
        Insert: {
          cardio_name: string;
          created_at?: string;
          duration_minutes: number;
          end_time: string;
          id?: string;
          start_time: string;
          user_email: string;
          user_id?: string | null;
        };
        Update: {
          cardio_name?: string;
          created_at?: string;
          duration_minutes?: number;
          end_time?: string;
          id?: string;
          start_time?: string;
          user_email?: string;
          user_id?: string | null;
        };
        Relationships: [];
      };
// ...
      daily_weights_exercises: {
        Row: {
          created_at: string;
          end_time: string | null;
          id: string;
          start_time: string;
          user_email: string;
          user_id: string | null;
        };
        Insert: {
          created_at?: string;
          end_time?: string | null;
          id?: string;
          start_time: string;
          user_email: string;
          user_id?: string | null;
        };
        Update: {
          created_at?: string;
          end_time?: string | null;
          id?: string;
          start_time?: string;
          user_email?: string;
          user_id?: string | null;
        };
        Relationships: [];
      };
      each_weights_exercises: {
        Row: {
          body_part: string;
          id: string;
          user_email: string;
          user_id: string | null;
          weights_exercises_id: string;
          workout_name: string;
        };
        Insert: {
          body_part: string;
          id?: string;
          user_email: string;
          user_id?: string | null;
          weights_exercises_id: string;
          workout_name: string;
        };
        Update: {
          body_part?: string;
          id?: string;
          user_email?: string;
          user_id?: string | null;
          weights_exercises_id?: string;
          workout_name?: string;
        };
        Relationships: [
          {
            foreignKeyName: "each_weights_exercises_weights_exercises_fkey";
            columns: ["weights_exercises_id"];
            isOneToOne: false;
            referencedRelation: "daily_weights_exercises";
            referencedColumns: ["id"];
          },
        ];
      };
      each_weights_exercises_set_info: {
        Row: {
          created_at: string;
          each_weights_exercises_id: string;
          id: string;
          kg: number;
          reps: number;
          set_number: number;
          user_email: string;
          user_id: string | null;
        };
        Insert: {
          created_at?: string;
          each_weights_exercises_id: string;
          id?: string;
          kg: number;
          reps: number;
          set_number: number;
          user_email: string;
          user_id?: string | null;
        };
        Update: {
          created_at?: string;
          each_weights_exercises_id?: string;
          id?: string;
          kg?: number;
          reps?: number;
          set_number?: number;
          user_email?: string;
          user_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "each_weights_exercises_set_info_weights_exercises_fkey";
            columns: ["each_weights_exercises_id"];
            isOneToOne: false;
            referencedRelation: "each_weights_exercises";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      add_daily_workouts: {
        Args: {
          body: Json;
        };
        Returns: undefined;
      };
      add_meals: {
        Args: {
          body: Json;
        };
        Returns: undefined;
      };
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};